Investigate memory hog
It turns out the problem is that the segmented meshes are extremely large.
a distance matrix of 40K vertices can easily use more RAM than I have on
this laptop. Decimating the meshes helped with the memory use.

However, there is also the time complexity problem of using Floyd-Warshall
algorithm. It can take over 30 min to process a single model, which is
impractical for a single preprocessing step.